code review changes, and some additional cleanup
