+++ src/stages/stage6/background_anim.c
@@ -13,13 +13,18 @@
 
 #include "global.h"
 #include "stageutils.h"
-#include "extra.h"
+
+void start_fall_over(void) { //troll
+	Stage6DrawData *draw_data = stage6_get_draw_data();
+	draw_data->fall_over.frames = global.frames;
+}
 
 void stage6_update(void) {
+	Stage6DrawData *draw_data = stage6_get_draw_data();
 	stage3d_update(&stage_3d_context);
 
-	if(fall_over) {
-		int t = global.frames - fall_over;
+	if(draw_data->fall_over.frames) {
+		int t = global.frames - draw_data->fall_over.frames;
 		TIMER(&t);
 
 		FROM_TO(0, 240, 1) {
@@ -28,13 +33,13 @@ void stage6_update(void) {
 		}
 
 		FROM_TO(150, 1000, 1) {
-			stage_3d_context.crot[0] -= 0.02*(global.frames-fall_over-150);
+			stage_3d_context.crot[0] -= 0.02*(global.frames-draw_data->fall_over.frames-150);
 			if(stage_3d_context.crot[0] < 0)
 				stage_3d_context.crot[0] = 0;
 		}
 
 		if(t >= 190)
-			stage_3d_context.cx[2] -= fmax(6, 0.05*(global.frames-fall_over-150));
+			stage_3d_context.cx[2] -= fmax(6, 0.05*(global.frames-draw_data->fall_over.frames-150));
 
 		FROM_TO(300, 470,1) {
 			stage_3d_context.cx[0] -= 0.01*cos(M_PI/180*stage_3d_context.crot[2]+M_PI/2)*_i;
