+++ src/stages/stage6/nonspells/boss_nonspell_4.c
@@ -9,7 +9,6 @@
 #include "taisei.h"
 
 #include "nonspells.h"
-#include "../elly.h"
 
 #include "global.h"
